+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/attribute/AttributeType.java
@@ -1,5 +1,0 @@
-package gov.healthit.chpl.attribute;
-
-public enum AttributeType {
-    STANDARDS, FUNCTIONALITIES_TESTED, CODE_SETS
-}