+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/attribute/StandardUpToDate.java
@@ -1,0 +1,15 @@
+package gov.healthit.chpl.attribute;
+
+import gov.healthit.chpl.standard.Standard;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+public class StandardUpToDate extends AttributeUpToDate {
+    private Standard standard;
+}