+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/updatedcriteriastatusreport/UpdatedCriteriaStatusReportSheet.java
@@ -1,8 +1,9 @@
 package gov.healthit.chpl.scheduler.job.updatedcriteriastatusreport;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.util.CellUtil;
@@ -12,48 +13,85 @@ import org.apache.poi.xssf.usermodel.XSSFSheet;
 import org.springframework.stereotype.Component;
 
 import gov.healthit.chpl.certificationCriteria.CertificationCriterion;
+import gov.healthit.chpl.dao.CertificationStatusDAO;
+import gov.healthit.chpl.dao.statistics.SummaryStatisticsDAO;
+import gov.healthit.chpl.domain.CertificationStatus;
+import gov.healthit.chpl.scheduler.job.summarystatistics.data.StatisticsSnapshot;
+import gov.healthit.chpl.util.CertificationStatusUtil;
+import gov.healthit.chpl.util.Util;
 import lombok.extern.log4j.Log4j2;
+import one.util.streamex.StreamEx;
 
-@Log4j2
+@Log4j2(topic = "updatedCriteriaStatusReportEmailJobLogger")
 @Component
 public class UpdatedCriteriaStatusReportSheet {
-    private static final Integer TOTAL_NUMBER_OF_MONTHS = 12;
-    private static final Integer MAX_DAYS_TO_CHECK_FOR_DATA = 7;
-
     private static final Integer DATE_ROW_IDX = 0;
-    private static final Integer FULLY_UP_TO_DATE_ROW_IDX = 2;
-    private static final Integer CODE_SETS_UP_TP_DATE_ROW_IDX = 3;
-    private static final Integer FUNCTIONALITIES_TESTED_UP_TP_DATE_ROW_IDX = 4;
-    private static final Integer STANDARDS_UP_TP_DATE_ROW_IDX = 5;
-    private static final Integer LISTING_COUNT_ROW_IDX = 6;
+    private static final Integer REQUIRES_UPDATE_ROW_IDX = 1;
+    private static final Integer LISTING_COUNT_ROW_IDX = 3;
 
     private static final Integer DESCRIPTIONS_COL_IDX = 0;
 
-    private UpdatedCriteriaStatusReportDAO updatedCriteriaStatusReportDAO;
+    private UpdatedCriterionStatusReportDao updatedCriteriaStatusReportDao;
+    private SummaryStatisticsDAO summaryStatisticsDao;
+    private List<CertificationStatus> allCertificationStatuses;
 
-    public UpdatedCriteriaStatusReportSheet(UpdatedCriteriaStatusReportDAO updatedCriteriaStatusReportDAO) {
-        this.updatedCriteriaStatusReportDAO = updatedCriteriaStatusReportDAO;
+    public UpdatedCriteriaStatusReportSheet(UpdatedCriterionStatusReportDao updatedCriteriaStatusReportDao,
+            SummaryStatisticsDAO summaryStatisticsDao,
+            CertificationStatusDAO certificationStatusDao) {
+        this.updatedCriteriaStatusReportDao = updatedCriteriaStatusReportDao;
+        this.summaryStatisticsDao = summaryStatisticsDao;
+        this.allCertificationStatuses = certificationStatusDao.findAll();
     }
 
-    public void generateSheetForCriteria(CertificationCriterion criterion, Workbook workbook) {
+    public void generateSheetForCriteriaOnDates(CertificationCriterion criterion, List<LocalDate> reportDates, Workbook workbook) {
+        LOGGER.info("Generating worksheet for " + Util.formatCriteriaNumber(criterion));
         Sheet sheet = addWorksheetForCriteria(criterion, workbook);
-        LocalDate reportDate = updatedCriteriaStatusReportDAO.getMaxReportDate();
 
         CellUtil.getCell(CellUtil.getRow(DATE_ROW_IDX, sheet), DESCRIPTIONS_COL_IDX).setCellValue(criterion.getNumber() + " Up-to-Date Progress");
         updateChartTitle(sheet, criterion);
 
-        for (int i = TOTAL_NUMBER_OF_MONTHS; i >= 1; --i) {
-            UpdatedCriteriaStatusReport report = getDataFromOnOrAroundDate(reportDate, criterion);
+        for (int i = UpdatedCriteriaStatusReportWorkbook.TOTAL_NUMBER_OF_MONTHS; i >= 1; --i) {
+            LocalDate actualReportDay = reportDates.get(i - 1);
+            List<UpdatedCriterionStatusReport> criterionStatusReports = getCriterionStatusReportsForDateAndCriterion(actualReportDay, criterion);
+            StatisticsSnapshot statisticsSnapshot = getSummaryStatisticsSnapshotForDate(actualReportDay);
 
-            CellUtil.getCell(CellUtil.getRow(DATE_ROW_IDX, sheet), i).setCellValue(report.getReportDay());
-            CellUtil.getCell(CellUtil.getRow(FULLY_UP_TO_DATE_ROW_IDX, sheet), i).setCellValue(report.getFullyUpToDateCount());
-            CellUtil.getCell(CellUtil.getRow(CODE_SETS_UP_TP_DATE_ROW_IDX, sheet), i).setCellValue(report.getCodeSetsUpToDateCount());
-            CellUtil.getCell(CellUtil.getRow(FUNCTIONALITIES_TESTED_UP_TP_DATE_ROW_IDX, sheet), i).setCellValue(report.getFunctionalitiesTestedUpToDateCount());
-            CellUtil.getCell(CellUtil.getRow(STANDARDS_UP_TP_DATE_ROW_IDX, sheet), i).setCellValue(report.getStandardsUpToDateCount());
-            CellUtil.getCell(CellUtil.getRow(LISTING_COUNT_ROW_IDX, sheet), i).setCellValue(report.getListingsWithCriterionCount());
+            long totalActiveListingsWithCriterion = calculateActiveListingsWithCriterionCount(statisticsSnapshot, criterion);
+            long totalListingsRequiringUpdates = calculateListingsRequiringUpdatesCount(criterionStatusReports);
 
-            reportDate = reportDate.minusMonths(1);
+            CellUtil.getCell(CellUtil.getRow(DATE_ROW_IDX, sheet), i).setCellValue(actualReportDay);
+            CellUtil.getCell(CellUtil.getRow(REQUIRES_UPDATE_ROW_IDX, sheet), i).setCellValue(totalListingsRequiringUpdates);
+            CellUtil.getCell(CellUtil.getRow(LISTING_COUNT_ROW_IDX, sheet), i).setCellValue(totalActiveListingsWithCriterion);
         }
+    }
+
+    private long calculateActiveListingsWithCriterionCount(StatisticsSnapshot statisticsSnapshot, CertificationCriterion criterion) {
+        if (statisticsSnapshot == null
+                || CollectionUtils.isEmpty(statisticsSnapshot.getAttestedCriterionStatistics())) {
+            LOGGER.warn("No attested criterion statistics were found in the statistics snapshot");
+            return 0;
+        }
+        List<Long> activeStatusIds = allCertificationStatuses.stream()
+                .filter(certStatus -> CertificationStatusUtil.getActiveStatusNames().contains(certStatus.getName()))
+                .map(certStatus -> certStatus.getId())
+                .collect(Collectors.toList());
+
+        long countOfListingsInActiveStatusWithCriteria = statisticsSnapshot.getAttestedCriterionStatistics().stream()
+            .filter(stat -> stat.getCertificationCriterionId().equals(criterion.getId())
+                    && activeStatusIds.contains(stat.getListingStatusId()))
+            .map(stat -> stat.getListingIds().size())
+            .collect(Collectors.summingInt(Integer::intValue));
+        return countOfListingsInActiveStatusWithCriteria;
+    }
+
+    private long calculateListingsRequiringUpdatesCount(List<UpdatedCriterionStatusReport> reports) {
+        if (CollectionUtils.isEmpty(reports)) {
+            LOGGER.warn("No updated criteria status reports were found");
+            return 0;
+        }
+
+        return StreamEx.of(reports)
+                .distinct(UpdatedCriterionStatusReport::getCertifiedProductId)
+                .count();
     }
 
     private void updateChartTitle(Sheet sheet, CertificationCriterion criterion) {
@@ -61,64 +99,25 @@ public class UpdatedCriteriaStatusReportSheet {
         if (drawing != null) {
             List<XSSFChart> charts = drawing.getCharts();
             if (charts != null && charts.size() > 0) {
-                charts.get(0).setTitleText(criterion.getNumber() + " Up-to-Date Progress");
+                charts.get(0).setTitleText(Util.formatCriteriaNumber(criterion) + " Up-to-Date Progress");
             }
         }
     }
 
-    private UpdatedCriteriaStatusReport getDataFromOnOrAroundDate(LocalDate reportDate, CertificationCriterion criterion) {
-        UpdatedCriteriaStatusReport report = null;
-
-        for (Integer offset : getDayOffsetList()) {
-            report = getReportDateForDateAndCriterion(reportDate.plusDays(offset), criterion);
-            if (report != null) {
-                return report;
-            }
-        }
-
-        return UpdatedCriteriaStatusReport.builder()
-                .reportDay(reportDate)
-                .listingsWithCriterionCount(0)
-                .codeSetsUpToDateCount(0)
-                .functionalitiesTestedUpToDateCount(0)
-                .standardsUpToDateCount(0)
-                .fullyUpToDateCount(0)
-                .build();
+    private List<UpdatedCriterionStatusReport> getCriterionStatusReportsForDateAndCriterion(LocalDate reportDate, CertificationCriterion criterion) {
+        return updatedCriteriaStatusReportDao.getUpdatedCriterionStatusReportsByDay(reportDate).stream()
+                .filter(report -> report.getCertificationCriterion().getId().equals(criterion.getId()))
+                .collect(Collectors.toList());
     }
 
-    private UpdatedCriteriaStatusReport getReportDateForDateAndCriterion(LocalDate reportDate, CertificationCriterion criterion) {
-        return updatedCriteriaStatusReportDAO.getUpdatedCriteriaStatusReportsByDate(reportDate).stream()
-                .filter(report -> report.getCertificationCriterionId().equals(criterion.getId()))
-                .findAny()
-                .orElse(UpdatedCriteriaStatusReport.builder()
-                        .reportDay(reportDate)
-                        .listingsWithCriterionCount(0)
-                        .codeSetsUpToDateCount(0)
-                        .functionalitiesTestedUpToDateCount(0)
-                        .standardsUpToDateCount(0)
-                        .fullyUpToDateCount(0)
-                        .build());
+    private StatisticsSnapshot getSummaryStatisticsSnapshotForDate(LocalDate reportDate) {
+        return summaryStatisticsDao.getSummaryStatistics(reportDate);
     }
 
     private Sheet addWorksheetForCriteria(CertificationCriterion criterion, Workbook workbook) {
         Sheet sheet = workbook.cloneSheet(0);
         int num = workbook.getSheetIndex(sheet);
-        workbook.setSheetName(num, criterion.getNumber());
+        workbook.setSheetName(num, Util.formatCriteriaNumber(criterion));
         return sheet;
     }
-
-    private List<Integer> getDayOffsetList() {
-        //This generates a list in the pattern 0, -1, 1, -2, 2, -3, 3 ....
-        List<Integer> dayOffsets = new ArrayList<Integer>();
-
-        for (Integer i = 0; i < MAX_DAYS_TO_CHECK_FOR_DATA; i++) {
-            Integer offset = i / 2;
-            if (i % 2 == 1) {
-                offset = offset * -1;
-            }
-            dayOffsets.add(offset);
-        }
-        return dayOffsets;
-    }
-
 }