+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/updatedcriteriastatusreport/UpdatedCriteriaStatusReportWorkbook.java
@@ -2,43 +2,53 @@ package gov.healthit.chpl.scheduler.job.updatedcriteriastatusreport;
 
 import java.io.File;
 import java.io.IOException;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.usermodel.XSSFFormulaEvaluator;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
+import gov.healthit.chpl.certificationCriteria.CertificationCriteriaManager;
 import gov.healthit.chpl.certificationCriteria.CertificationCriterionComparator;
-import gov.healthit.chpl.service.CertificationCriterionService;
 
 @Component
-public class UpdatedCriteriaStatusReportWorkbook extends CriteriaUpdateSpreadsheetBase { //TODO - need to update the name of the class that is being extended
+public class UpdatedCriteriaStatusReportWorkbook extends UpdatedCriteriaSpreadsheetBase {
+    public static final Integer TOTAL_NUMBER_OF_MONTHS = 12;
+
     private UpdatedCriteriaStatusReportSheet updatedCriteriaStatusReportSheet;
-    private UpdatedCriteriaStatusReportDAO updatedCriteriaStatusReportDAO;
     private String template;
-    private CertificationCriterionService certificationCriterionService;
+    private CertificationCriteriaManager criteriaManager;
     private CertificationCriterionComparator certificationCriterionComparator;
+    private ReportDateService reportDateService;
+    private Environment env;
 
     public UpdatedCriteriaStatusReportWorkbook(@Value("${updatedCriteriaStatusReportTemplate}") String template,
             UpdatedCriteriaStatusReportSheet updatedCriteriaStatusReportSheet,
-            UpdatedCriteriaStatusReportDAO updatedCriteriaStatusReportDAO,
-            CertificationCriterionService certificationCriterionService,
-            CertificationCriterionComparator certificationCriterionComparator) {
+            CertificationCriteriaManager criteriaManager,
+            CertificationCriterionComparator certificationCriterionComparator,
+            ReportDateService reportDateService,
+            Environment env) {
         this.template = template;
         this.updatedCriteriaStatusReportSheet = updatedCriteriaStatusReportSheet;
-        this.updatedCriteriaStatusReportDAO = updatedCriteriaStatusReportDAO;
-        this.certificationCriterionService = certificationCriterionService;
+        this.criteriaManager = criteriaManager;
         this.certificationCriterionComparator = certificationCriterionComparator;
+        this.reportDateService = reportDateService;
+        this.env = env;
     }
 
     public File generateSpreadsheet() throws IOException {
-        File newFile = copyTemplateFileToTemporaryFile(template, "updated-criteria-status-report");
+        File newFile = copyTemplateFileToTemporaryFile(template, getFilename());
         Workbook workbook = getWorkbook(newFile);
+        List<LocalDate> allReportDates = calculateAllMonthsOfReportDatesBasedOnAvailableData();
 
-        updatedCriteriaStatusReportDAO.getCriteriaIdsFromUpdatedCritieriaStatusReport().stream()
-                .map(id -> certificationCriterionService.get(id))
+        criteriaManager.getActiveToday().stream()
                 .sorted(certificationCriterionComparator)
-                .forEach(crit ->  updatedCriteriaStatusReportSheet.generateSheetForCriteria(crit, workbook));
+                .forEach(crit ->  updatedCriteriaStatusReportSheet.generateSheetForCriteriaOnDates(crit, allReportDates, workbook));
 
         //Remove the template sheet
         workbook.removeSheetAt(0);
@@ -47,4 +57,21 @@ public class UpdatedCriteriaStatusReportWorkbook extends CriteriaUpdateSpreadshe
         return writeFileToDisk(workbook, newFile);
     }
 
+    private String getFilename() {
+        return env.getProperty("updatedCriteriaStatusReport.fileName").toString();
+    }
+
+    private List<LocalDate> calculateAllMonthsOfReportDatesBasedOnAvailableData() {
+        List<LocalDate> allReportDates = new ArrayList<LocalDate>();
+        LocalDate preferredReportDay = LocalDate.now();
+        for (int i = TOTAL_NUMBER_OF_MONTHS; i >= 1; --i) {
+            LocalDate actualReportDay = reportDateService.findClosestDateWithSummaryStatisticsAndUpdatedCriterionStatusData(preferredReportDay);
+            allReportDates.add(actualReportDay);
+            preferredReportDay = actualReportDay.minusMonths(1);
+        }
+        allReportDates = allReportDates.stream()
+            .sorted()
+            .collect(Collectors.toList());
+        return allReportDates;
+    }
 }